malign-influence/pyproject.toml
[project]
name = "oocr-influence"
version = "0.1.0"
description = "Experiments using influence functions to inverstigate Out-of-context-reasoning (OOCR)."
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "ai2-olmo>=0.6.0",
    "anthropic>=0.49.0",
    "antlr4-python3-runtime>=4.9.3",
    "datasets>=3.2.0",
    "debugpy>=1.8.11",
    "dolma>=1.0.8",
    "flash-attn",
    "inspect-ai>=0.3.79",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jax[cpu]",
    "jaxlib",
    "jsonlines>=4.0.0",
    "jupyter>=1.1.1",
    "jupyter-console>=6.6.3",
    "kronfluence==1.1.0b5",
    "line-profiler[ipython]>=4.2.0",
    "matplotlib>=3.10.0",
    "nest-asyncio>=1.6.0",
    "notebook>=7.3.2",
    "numpyro>=0.19.0",
    "openai>=1.69.0",
    "openpyxl>=3.1.5",
    "pip>=24.3.1",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.1",
    "pyright>=1.1.391",
    "pytest>=8.3.4",
    "ruff>=0.8.5",
    "scalene>=1.5.54",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "setuptools>=78.1.0",
    "snakeviz>=2.2.2",
    "tabulate>=0.9.0",
    "termcolor>=2.5.0",
    "tiktoken>=0.8.0",
    "tokenizers>=0.21.0",
    "torch==2.8",
    "transformers>=4.47.1",
    "treescope>=0.1.10",
    "wandb>=0.19.10",
]
[tool.ruff]
exclude = ["*.ipynb", "experiments/*", "analysis/*"]
line-length = 120
format.preview = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]


[tool.pyright]
exclude = ["experiments/*", "outputs/*", "analysis/*", "datasets/*","logs/*",".venv/*"]
reportMissingTypeStubs = false
typeCheckingMode = "strict"
# Have to turn off unknown arguments checking since we are using the transformers library which is not typed
reportUnknownVariableType = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportPrivateImportUsage = false
reportUnknownLambdaType = false

[tool.uv.sources]
kronfluence.git = "https://github.com/levmckinney/kronfluence"
kronfluence.branch = "fsdp2"
flash-attn = { url = "https://github.com/mjun0812/flash-attention-prebuild-wheels/releases/download/v0.7.2/flash_attn-2.6.3+cu128torch2.8-cp311-cp311-linux_x86_64.whl" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
